Move connector helpers into connectors crate
