+++ codex-rs/chatgpt/src/connectors.rs
@@ -1,7 +1,3 @@
-use codex_core::config::Config;
-use codex_login::AuthManager;
-use codex_login::CodexAuth;
-use codex_login::token_data::TokenData;
 use std::collections::HashSet;
 use std::time::Duration;
 
@@ -9,21 +5,24 @@ use crate::chatgpt_client::chatgpt_get_request_with_timeout;
 use crate::chatgpt_token::get_chatgpt_token_data;
 use crate::chatgpt_token::init_chatgpt_token_from_auth;
 
+use codex_app_server_protocol::AppInfo;
 use codex_connectors::AllConnectorsCacheKey;
 use codex_connectors::DirectoryListResponse;
-
-pub use codex_core::connectors::AppInfo;
-pub use codex_core::connectors::connector_display_label;
-use codex_core::connectors::filter_disallowed_connectors;
+use codex_connectors::filter::filter_disallowed_connectors;
+use codex_connectors::merge::merge_connectors;
+use codex_connectors::merge::merge_plugin_connectors;
+use codex_core::config::Config;
 pub use codex_core::connectors::list_accessible_connectors_from_mcp_tools;
 pub use codex_core::connectors::list_accessible_connectors_from_mcp_tools_with_options;
 pub use codex_core::connectors::list_accessible_connectors_from_mcp_tools_with_options_and_status;
 pub use codex_core::connectors::list_cached_accessible_connectors_from_mcp_tools;
-use codex_core::connectors::merge_connectors;
-use codex_core::connectors::merge_plugin_apps;
 pub use codex_core::connectors::with_app_enabled_state;
 use codex_core::plugins::AppConnectorId;
 use codex_core::plugins::PluginsManager;
+use codex_login::AuthManager;
+use codex_login::CodexAuth;
+use codex_login::default_client::originator;
+use codex_login::token_data::TokenData;
 
 const DIRECTORY_CONNECTORS_TIMEOUT: Duration = Duration::from_secs(60);
 
@@ -74,8 +73,17 @@ pub async fn list_cached_all_connectors(config: &Config) -> Option<Vec<AppInfo>>
     let token_data = get_chatgpt_token_data()?;
     let cache_key = all_connectors_cache_key(config, &token_data);
     let connectors = codex_connectors::cached_all_connectors(&cache_key)?;
-    let connectors = merge_plugin_apps(connectors, plugin_apps_for_config(config).await);
-    Some(filter_disallowed_connectors(connectors))
+    let connectors = merge_plugin_connectors(
+        connectors,
+        plugin_apps_for_config(config)
+            .await
+            .into_iter()
+            .map(|connector_id| connector_id.0),
+    );
+    Some(filter_disallowed_connectors(
+        connectors,
+        originator().value.as_str(),
+    ))
 }
 
 pub async fn list_all_connectors_with_options(
@@ -105,8 +113,17 @@ pub async fn list_all_connectors_with_options(
         },
     )
     .await?;
-    let connectors = merge_plugin_apps(connectors, plugin_apps_for_config(config).await);
-    Ok(filter_disallowed_connectors(connectors))
+    let connectors = merge_plugin_connectors(
+        connectors,
+        plugin_apps_for_config(config)
+            .await
+            .into_iter()
+            .map(|connector_id| connector_id.0),
+    );
+    Ok(filter_disallowed_connectors(
+        connectors,
+        originator().value.as_str(),
+    ))
 }
 
 fn all_connectors_cache_key(config: &Config, token_data: &TokenData) -> AllConnectorsCacheKey {
@@ -134,7 +151,13 @@ pub fn connectors_for_plugin_apps(
         .map(|connector_id| connector_id.0.as_str())
         .collect::<HashSet<_>>();
 
-    filter_disallowed_connectors(merge_plugin_apps(connectors, plugin_apps.to_vec()))
+    let connectors = merge_plugin_connectors(
+        connectors,
+        plugin_apps
+            .iter()
+            .map(|connector_id| connector_id.0.clone()),
+    );
+    filter_disallowed_connectors(connectors, originator().value.as_str())
         .into_iter()
         .filter(|connector| plugin_app_ids.contains(connector.id.as_str()))
         .collect()
@@ -158,13 +181,13 @@ pub fn merge_connectors_with_accessible(
         accessible_connectors
     };
     let merged = merge_connectors(connectors, accessible_connectors);
-    filter_disallowed_connectors(merged)
+    filter_disallowed_connectors(merged, originator().value.as_str())
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use codex_core::connectors::connector_install_url;
+    use codex_connectors::metadata::connector_install_url;
     use codex_core::plugins::AppConnectorId;
     use pretty_assertions::assert_eq;
 
@@ -184,46 +207,6 @@ mod tests {
             is_enabled: true,
             plugin_display_names: Vec::new(),
         }
-    }
-
-    #[test]
-    fn allows_asdk_connectors() {
-        let filtered = filter_disallowed_connectors(vec![app("asdk_app_hidden"), app("alpha")]);
-        assert_eq!(filtered, vec![app("asdk_app_hidden"), app("alpha")]);
-    }
-
-    #[test]
-    fn allows_whitelisted_asdk_connectors() {
-        let filtered = filter_disallowed_connectors(vec![
-            app("asdk_app_69781557cc1481919cf5e9824fa2e792"),
-            app("beta"),
-        ]);
-        assert_eq!(
-            filtered,
-            vec![
-                app("asdk_app_69781557cc1481919cf5e9824fa2e792"),
-                app("beta")
-            ]
-        );
-    }
-
-    #[test]
-    fn filters_openai_prefixed_connectors() {
-        let filtered = filter_disallowed_connectors(vec![
-            app("connector_openai_foo"),
-            app("connector_openai_bar"),
-            app("gamma"),
-        ]);
-        assert_eq!(filtered, vec![app("gamma")]);
-    }
-
-    #[test]
-    fn filters_disallowed_connector_ids() {
-        let filtered = filter_disallowed_connectors(vec![
-            app("asdk_app_6938a94a61d881918ef32cb999ff937c"),
-            app("delta"),
-        ]);
-        assert_eq!(filtered, vec![app("delta")]);
     }
 
     fn merged_app(id: &str, is_accessible: bool) -> AppInfo {
